pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-tools-converter"
version = "0.1.0"
description = "Convert Claude Code and OpenCode skills, agents, and commands between formats"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
agent-tools-converter = "agent_tools_converter.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
include = ["agent_tools_converter*"]
